Move stories to a sperate folder
