+++ packages/extension/src/stories/args.ts
@@ -1,0 +1,6 @@
+export const NETWORK_ARGTYPE = {
+  options: ['Mainnet', 'Ropsten', 'Rinkeby', 'Goerli', 'Kovan', 'xDai', 'Localhost', 'Hardhat', 'Other'],
+  control: {
+    type: 'select',
+  },
+}
