+++ packages/extension/src/stories/components/EventPreview/FetchError.stories.tsx
@@ -2,9 +2,9 @@ import React from 'react'
 import type { Meta, Story } from '@storybook/react'
 import { Interface } from '@ethersproject/abi'
 
-import { GlobalStyle } from '../../../../providers/GlobalStyle'
-import { FetchErrorPreview } from '../FetchErrorPreview'
-import { AbiProvider } from '../../../../providers/abi/AbiProvider'
+import { GlobalStyle } from '../../../providers/GlobalStyle'
+import { FetchErrorPreview } from '../../../views/Events/EventPreview/FetchErrorPreview'
+import { AbiProvider } from '../../../providers/abi/AbiProvider'
 
 export default {
   title: 'Components/EventPreview/Fetch Error',
