+++ packages/extension/src/stories/components/EventPreview/InitializedPreview.stories.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import type { Story, Meta } from '@storybook/react'
 
-import { GlobalStyle } from '../../../../providers/GlobalStyle'
-import { InitializedPreview } from '../InitializedPreview'
+import { GlobalStyle } from '../../../providers/GlobalStyle'
+import { InitializedPreview } from '../../../views/Events/EventPreview/InitializedPreview'
 
 export default {
   title: 'Components/EventPreview/Initialized',
