+++ packages/extension/src/stories/pages/Events.stories.tsx
@@ -4,7 +4,7 @@ import { Interface } from '@ethersproject/abi'
 
 import { GlobalStyle } from '../../providers/GlobalStyle'
 import { EventContext } from '../../providers/events/EventProvider'
-import { Events as EventsComponent } from './Events'
+import { Events as EventsComponent } from '../../views/Events/Events'
 import type { Event } from '../../providers/events/State'
 import { DEFAULT_ABIS } from '../../providers/abi/defaultAbis'
 import { AbiProvider } from '../../providers/abi/AbiProvider'
